requirements.txt
dnspython>=2.4.0
aiohttp>=3.9.0
rich>=13.0.0
